Added needed packages to Pipfile, created a tentative function for creating veteran counts visualization
