+++ .gitignore
@@ -144,3 +144,6 @@ cython_debug/
 .elasticbeanstalk/*
 !.elasticbeanstalk/*.cfg.yml
 !.elasticbeanstalk/*.global.yml
+
+# Practice files
+practice_connection.py
