+++ app/main.py
@@ -24,7 +24,7 @@ app = FastAPI(
 
 app.include_router(db.router, tags=['Database'])
 app.include_router(ml.router, tags=['Machine Learning'])
-app.include_router(viz.router, tags=['Visualization'])
+app.include_router(viz.router, tags=['Visualizations'])
 
 app.add_middleware(
     CORSMiddleware,
